pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "147template"
version = "0.1.0"
[tool.poetry.dependencies]
Pillow = "^7.0"
python = "^3.7"
ansicolors = "^1.1"
[tool.poetry.dev-dependencies]
